+++ crates/zizmor/src/audit/mod.rs
@@ -14,7 +14,7 @@ use crate::{
         AsDocument,
         action::{Action, CompositeStep, DockerAction},
         dependabot::Dependabot,
-        pre_commit::{PreCommitConfig, PreCommitHooks},
+        pre_commit::{self, PreCommitConfig, PreCommitHooks},
         workflow::{Job, NormalJob, ReusableWorkflowCallJob, Step, Workflow},
     },
     registry::input::InputKey,
@@ -39,6 +39,7 @@ pub(crate) mod github_env;
 pub(crate) mod hardcoded_container_credentials;
 pub(crate) mod impostor_commit;
 pub(crate) mod insecure_commands;
+pub(crate) mod insecure_url_scheme;
 pub(crate) mod known_vulnerable_actions;
 pub(crate) mod misfeature;
 pub(crate) mod obfuscation;
@@ -100,6 +101,18 @@ impl AuditInput {
             AuditInput::PreCommitConfig(pre_commit_config) => pre_commit_config.location(),
             AuditInput::PreCommitHooks(pre_commit_hooks) => pre_commit_hooks.location(),
         }
+    }
+
+    /// Returns whether this kind of input supports GitHub Actions' template syntax,
+    /// i.e. "actions expressions."
+    ///
+    /// This exists because some [`Audit::audit_raw`] implementations exist to walk
+    /// actions expressions, but not all raw inputs can actually contain those expressions.
+    ///
+    /// TODO: This is kind of goofy. Maybe we should do this by construction,
+    /// i.e. have an `Audit::audit_raw_gha` instead.
+    pub(crate) fn supports_gha_template_syntax(&self) -> bool {
+        matches!(self, AuditInput::Workflow(_) | AuditInput::Action(_))
     }
 }
 
@@ -258,7 +271,7 @@ impl AuditError {
 /// Auditing trait.
 ///
 /// Implementors of this trait can choose the level of specificity/context
-/// they need for workflows and/or action definitions:
+/// they need for their kind(s) of input:
 ///
 /// For workflows:
 ///
@@ -269,11 +282,18 @@ impl AuditError {
 ///
 /// For actions:
 ///
-/// 1. [`Audit::audit_action`]: runs at the top of the action (most general)
-/// 2. [`Audit::audit_composite_step`]: runs on each composite step within the
+/// 1. [`Audit::audit_docker_action`]: runs at the top of the Docker action (most general)
+/// 1. [`Audit::audit_action`]: runs at the top of the composite action (most general)
+/// 1. [`Audit::audit_composite_step`]: runs on each composite step within the
 ///    action (most specific)
 ///
-/// For both:
+/// For pre-commit inputs:
+///
+/// 1. [`Audit::audit_pre_commit_config`]: runs at the top of the pre-commit configuration (most general)
+/// 1. [`Audit::audit_pre_commit_config_repo`]: runs on each `repo` definition within the pre-commit
+///    configuration
+///
+/// For all:
 ///
 /// 1. [`Audit::audit_raw`]: runs on the raw, unparsed YAML document source
 ///
@@ -380,6 +400,28 @@ pub(crate) trait Audit: AuditCore {
         Ok(vec![])
     }
 
+    async fn audit_pre_commit_config_repo<'doc>(
+        &self,
+        _repo: &pre_commit::Repo<'doc>,
+        _config: &Config,
+    ) -> Result<Vec<Finding<'doc>>, AuditError> {
+        Ok(vec![])
+    }
+
+    async fn audit_pre_commit_config<'doc>(
+        &self,
+        pre_commit: &'doc PreCommitConfig,
+        config: &Config,
+    ) -> Result<Vec<Finding<'doc>>, AuditError> {
+        let mut results = vec![];
+
+        for repo in pre_commit.repos() {
+            results.extend(self.audit_pre_commit_config_repo(&repo, config).await?);
+        }
+
+        Ok(results)
+    }
+
     async fn audit_raw<'doc>(
         &self,
         _input: &'doc AuditInput,
@@ -388,7 +430,7 @@ pub(crate) trait Audit: AuditCore {
         Ok(vec![])
     }
 
-    /// The top-level auditing function for both workflows and actions.
+    /// The top-level auditing function for all inputs.
     ///
     /// Implementors **should not** override this blanket implementation,
     /// since it's marked with tracing instrumentation.
@@ -419,8 +461,11 @@ pub(crate) trait Audit: AuditCore {
             AuditInput::Workflow(workflow) => self.audit_workflow(workflow, config).await,
             AuditInput::Action(action) => self.audit_action(action, config).await,
             AuditInput::Dependabot(dependabot) => self.audit_dependabot(dependabot, config).await,
-            AuditInput::PreCommitConfig(_) | AuditInput::PreCommitHooks(_) => {
-                warn_once!("pre-commit auditing not implemented yet");
+            AuditInput::PreCommitConfig(pre_commit) => {
+                self.audit_pre_commit_config(pre_commit, config).await
+            }
+            AuditInput::PreCommitHooks(_) => {
+                warn_once!("pre-commit hooks auditing not implemented yet");
                 Ok(vec![])
             }
         }?;