+++ crates/zizmor/src/audit/overprovisioned_secrets.rs
@@ -41,6 +41,10 @@ impl Audit for OverprovisionedSecrets {
     ) -> Result<Vec<super::Finding<'doc>>, AuditError> {
         let mut findings = vec![];
 
+        if !input.supports_gha_template_syntax() {
+            return Ok(findings);
+        }
+
         for (expr, span) in parse_fenced_expressions_from_routable(input) {
             let Ok(parsed) = Expr::parse(expr.as_bare()) else {
                 tracing::warn!("couldn't parse expression: {expr}", expr = expr.as_bare());